% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barconfph.R
\name{barconfPh}
\alias{barconfPh}
\title{Barplot simple avec les IC a partir d'une donnee discrete.}
\usage{
barconfPh(dfx, lig, titre = "", titx = "", tity = "\%", seuil = 95,
  desc = TRUE, angle = FALSE)
}
\arguments{
\item{dfx}{un data.frame}

\item{lig}{une variable factorielle}

\item{titre}{Titre du graphique (vide par defaut)}

\item{titx}{Titre de l'axe des x (vide par defaut)}

\item{tity}{Titre de l'axe des y (pourcent par defaut)}

\item{seuil}{Marge d'erreur en % (95 par défaut)}

\item{desc}{Classification descendante des valeurs (TRUE par default)}

\item{angle}{Niveaux en x inscrit avec un angle (FALSE par défaut)}
}
\value{
un graphique
}
\description{
Trace un graphique en barre avec les IC à 95 % des  pourcentages d'une variable fcatorielle.
Calcul des intervalles de confiance par transformation angulaire.
}
