% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsimple.R
\name{corsimpleph}
\alias{corsimpleph}
\title{Graphique de correlation simple}
\usage{
corsimpleph(
  dfx,
  xx,
  yy,
  titre = "",
  titx = "",
  tity = "",
  moymob = TRUE,
  dtcor = FALSE
)
}
\arguments{
\item{dfx}{Data frame à étudier}

\item{xx}{Variable en x}

\item{yy}{Variabe en y}

\item{titre}{Titre Principal}

\item{titx}{Titre de l'axe des X}

\item{tity}{Titre de l'axe des Y}

\item{moymob}{Trace la moyenne mobile + IC (TRUE par defaut)}

\item{dtcor}{Trace la droite de regression + IC (FALSE par defaut)}
}
\value{
un graphique
}
\description{
Trace le graphique en point avec la courbe de moyene mobile ou la droite de regression, affiche le p.
}
\examples{
corsimpleph(iris,Sepal.Length,Petal.Length, titx="Pétales",
                      tity ="Pistil", titre = "Grand titre")

}
