% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densite.R
\name{densitePh}
\alias{densitePh}
\title{Courbe de Densité}
\usage{
densitePh(dfx, varx, vary = "1", titre = "", titx = "")
}
\arguments{
\item{dfx}{Data.frame}

\item{varx}{Variable numérique}

\item{vary}{Groupes à représenter ("1" par défaut, un seul groupe)}

\item{titre}{titre du graphique ("" par défaut)}

\item{titx}{titre de l'axe des x ("" par defaut)}
}
\value{
graphique
}
\description{
Courbe de Densité
}
