% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxph.R
\name{graphboxph}
\alias{graphboxph}
\title{Box-plot}
\usage{
graphboxph(dfx, xx, yy, titre = "", titx = "", tity = "")
}
\arguments{
\item{dfx}{Data.frame}

\item{xx}{valeur catégorielle}

\item{yy}{valeur numérique}

\item{titre}{Titre principal (vide par defaut)}

\item{titx}{Titre de l'axe des x (vide par defaut)}

\item{tity}{Titre de l'axe des y (vide par defaut)}
}
\value{
graphique
}
\description{
Box-plot
}
