% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{histoPh}
\alias{histoPh}
\title{Histogramme}
\usage{
histoPh(dfx, varz, taille = 10, titre = "", titx = "", tity = "n")
}
\arguments{
\item{dfx}{Data.frame}

\item{varz}{Variable numérique}

\item{taille}{Largeur de chaque colonne (10 par défaut)}

\item{titre}{Titre du graphique ("" par défaut)}

\item{titx}{Titre des x ("" par défaut)}

\item{tity}{Titre des y ("n" par défaut)}
}
\value{
graphique
}
\description{
Histogramme
}
