% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramide.R
\name{pyramidph}
\alias{pyramidph}
\title{Pyramide des âges}
\usage{
pyramidph(dfx, age, sexe)
}
\arguments{
\item{dfx}{data.frame}

\item{age}{variable age "age" par défaut}

\item{sexe}{variable "sexe" sexe par défaut}
}
\value{
graphique
}
\description{
Pyramide des âges
}
