% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tang.R
\name{tangangPh}
\alias{tangangPh}
\title{Calcul des IC par transformation angulaire}
\usage{
tangangPh(nn, tot, seuil = 95)
}
\arguments{
\item{nn}{nb de cas}

\item{tot}{total}

\item{seuil}{: marge d'erreur exprimée en % (95 % par défaut)}
}
\value{
un  vecteur : pourcentage, borne inf, borne sup
}
\description{
Calcul des IC par transformation angulaire
}
